% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_transfer_entropy.R
\name{bootstrap_qte}
\alias{bootstrap_qte}
\title{Bootstrap QTE Confidence Intervals}
\usage{
bootstrap_qte(x, y, tau = 0.5, B = 1000, block_size = 20, conf_level = 0.95)
}
\arguments{
\item{x}{Source time series}

\item{y}{Target time series}

\item{tau}{Quantile level}

\item{B}{Number of bootstrap replications}

\item{block_size}{Block size for block bootstrap}

\item{conf_level}{Confidence level for intervals}
}
\value{
List containing bootstrap samples and confidence intervals
}
\description{
Bootstrap QTE Confidence Intervals
}
\examples{
\dontrun{
x <- rnorm(100)
y <- 0.5 * x + rnorm(100, sd = 0.1)
boot_results <- bootstrap_qte(x, y, tau = 0.5, B = 1000)
}
}
