% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagion_analysis.R
\name{calculate_contagion_index}
\alias{calculate_contagion_index}
\title{Calculate Contagion Index}
\usage{
calculate_contagion_index(returns_dm, returns_em, tau = 0.05, n.levels = 4)
}
\arguments{
\item{returns_dm}{Matrix of developed market returns}

\item{returns_em}{Matrix of emerging market returns}

\item{tau}{Quantile level (default: 0.05)}

\item{n.levels}{Number of wavelet levels (default: 4)}
}
\value{
Array containing contagion measures
}
\description{
Calculate Contagion Index
}
\examples{
\dontrun{
dm_returns <- get_stock_data(c("SPY", "EWU"))
em_returns <- get_stock_data(c("EWZ", "FXI"))
contagion <- calculate_contagion_index(dm_returns, em_returns)
}
}
