% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagion_analysis.R
\name{calculate_dynamic_contagion}
\alias{calculate_dynamic_contagion}
\title{Calculate Dynamic Contagion Index}
\usage{
calculate_dynamic_contagion(
  returns_dm,
  returns_em,
  window_size = 252,
  tau = 0.05
)
}
\arguments{
\item{returns_dm}{Matrix of developed market returns}

\item{returns_em}{Matrix of emerging market returns}

\item{window_size}{Rolling window size (default: 252)}

\item{tau}{Quantile level (default: 0.05)}
}
\value{
Matrix of dynamic contagion indices
}
\description{
Calculate Dynamic Contagion Index
}
\examples{
\donttest{
dm_returns <- get_stock_data(c("SPY", "EWU"))
em_returns <- get_stock_data(c("EWZ", "FXI"))
dci <- calculate_dynamic_contagion(dm_returns, em_returns)
}
}
