% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{calculate_network_metrics}
\alias{calculate_network_metrics}
\title{Calculate Network Metrics}
\usage{
calculate_network_metrics(network)
}
\arguments{
\item{network}{Network object from create_qte_network}
}
\value{
List of network metrics
}
\description{
Calculate Network Metrics
}
\examples{
\donttest{
data <- matrix(rnorm(300), ncol = 3)
wave <- lapply(1:ncol(data), function(i) wavelet_decompose(data[,i]))
net <- create_qte_network(data, wave, scale = 1)
metrics <- calculate_network_metrics(net)
}
}
