% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_transfer_entropy.R
\name{calculate_qte}
\alias{calculate_qte}
\title{Calculate Quantile Transfer Entropy}
\usage{
calculate_qte(x, y, tau = 0.5, k = 1)
}
\arguments{
\item{x}{Source time series}

\item{y}{Target time series}

\item{tau}{Quantile level (default: 0.5)}

\item{k}{Lag order (default: 1)}
}
\value{
Numeric QTE value or NA if calculation fails
}
\description{
Calculate Quantile Transfer Entropy
}
\examples{
\donttest{
x <- rnorm(100)
y <- 0.5 * x + rnorm(100, sd = 0.1)
qte <- calculate_qte(x, y, tau = 0.5)
}
}
