% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{calculate_summary_stats}
\alias{calculate_summary_stats}
\title{Calculate Summary Statistics}
\usage{
calculate_summary_stats(returns)
}
\arguments{
\item{returns}{xts object containing return data}
}
\value{
List of statistics
}
\description{
Calculate Summary Statistics
}
\examples{
\donttest{
data <- get_stock_data(data_source = "builtin")
stats <- calculate_summary_stats(data)
}
}
