% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{check_data_quality}
\alias{check_data_quality}
\title{Check Data Quality}
\usage{
check_data_quality(market_data, crisis_periods = NULL)
}
\arguments{
\item{market_data}{List of market data}

\item{crisis_periods}{List of crisis period dates (optional)}
}
\value{
NULL (prints diagnostics)
}
\description{
Check Data Quality
}
\examples{
\dontrun{
data <- list(developed = get_stock_data("AAPL"))
crisis_periods <- list(covid = c("2020-01-01", "2020-12-31"))
check_data_quality(data, crisis_periods)
}
\dontrun{
# Using built-in dataset
data <- get_stock_data(data_source = "builtin")
processed_data <- process_returns(data)
head(processed_data)
}
}
