% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{create_multiscale_networks}
\alias{create_multiscale_networks}
\title{Create Multiscale Networks}
\usage{
create_multiscale_networks(returns, wave_decomp, params)
}
\arguments{
\item{returns}{Return data matrix}

\item{wave_decomp}{Wavelet decomposition list}

\item{params}{List with n_levels and tau_levels}
}
\value{
List of networks and metrics
}
\description{
Create Multiscale Networks
}
\examples{
\donttest{
data <- matrix(rnorm(300), ncol = 3)
wave <- lapply(1:ncol(data), function(i) wavelet_decompose(data[,i]))
params <- list(n_levels = 4, tau_levels = c(0.1, 0.5, 0.9))
networks <- create_multiscale_networks(data, wave, params)
}
}
