% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_analysis.R
\name{create_qte_network}
\alias{create_qte_network}
\title{Create QTE Network}
\usage{
create_qte_network(returns, wave_decomp, scale, tau = 0.5)
}
\arguments{
\item{returns}{Return data matrix}

\item{wave_decomp}{List of wavelet decompositions}

\item{scale}{Scale level to analyze}

\item{tau}{Quantile level (default: 0.5)}
}
\value{
List containing network objects and QTE matrix
}
\description{
Create QTE Network
}
\examples{
\donttest{
data <- matrix(rnorm(300), ncol = 3)
wave <- lapply(1:ncol(data), function(i) wavelet_decompose(data[,i]))
net <- create_qte_network(data, wave, scale = 1)
}
}
