% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_results}
\alias{load_results}
\title{Load Analysis Results}
\usage{
load_results(file_path)
}
\arguments{
\item{file_path}{Path to saved results}
}
\value{
List of analysis results
}
\description{
Load Analysis Results
}
\examples{
\dontrun{
results <- load_results("analysis_results.RData")
}
}
