% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_enhanced_network}
\alias{plot_enhanced_network}
\title{Plot Enhanced Network}
\usage{
plot_enhanced_network(network, scale, tau)
}
\arguments{
\item{network}{Network object}

\item{scale}{Scale level}

\item{tau}{Quantile level}
}
\value{
ggplot object
}
\description{
Plot Enhanced Network
}
\examples{
\donttest{
net <- create_qte_network(returns, wave_decomp, scale = 1)
plot_enhanced_network(net, scale = 1, tau = 0.5)
}
}
