% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_market_strength}
\alias{plot_market_strength}
\title{Plot Market Strength}
\usage{
plot_market_strength(network, scale, tau)
}
\arguments{
\item{network}{Network object}

\item{scale}{Scale level}

\item{tau}{Quantile level}
}
\value{
ggplot object
}
\description{
Plot Market Strength
}
\examples{
\dontrun{
net <- create_qte_network(returns, wave_decomp, scale = 1)
plot_market_strength(net, scale = 1, tau = 0.5)
}
}
