% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_qte_heatmap}
\alias{plot_qte_heatmap}
\title{Plot QTE Heatmap}
\usage{
plot_qte_heatmap(qte_matrix, title = "Multiscale QTE Analysis")
}
\arguments{
\item{qte_matrix}{QTE results matrix}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot QTE Heatmap
}
\examples{
\donttest{
qte_results <- calculate_multiscale_qte(wave_x, wave_y, tau_levels)
plot_qte_heatmap(qte_results)
}
}
