% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_rolling_qte}
\alias{plot_rolling_qte}
\title{Plot Rolling QTE}
\usage{
plot_rolling_qte(rolling_qte, window_size)
}
\arguments{
\item{rolling_qte}{Rolling QTE results}

\item{window_size}{Window size used}
}
\value{
ggplot object
}
\description{
Plot Rolling QTE
}
\examples{
\dontrun{
rolling_results <- calculate_rolling_qte(x, y, window_size = 100)
plot_rolling_qte(rolling_results, window_size = 100)
}
}
