% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_spillover_heatmap}
\alias{plot_spillover_heatmap}
\title{Plot Spillover Heatmap}
\usage{
plot_spillover_heatmap(spillover_matrix, title = "Spillover Analysis")
}
\arguments{
\item{spillover_matrix}{Matrix of spillover values}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot Spillover Heatmap
}
\examples{
\donttest{
spillover_plot <- plot_spillover_heatmap(spillover_matrix)
}
}
