% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{save_plot}
\alias{save_plot}
\title{Save Plot to File}
\usage{
save_plot(plot, filename, width = 10, height = 8, dpi = 300)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{Output filename}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}

\item{dpi}{Resolution in dots per inch}
}
\value{
Invisible NULL
}
\description{
Save Plot to File
}
\examples{
\dontrun{
p <- plot_qte_heatmap(qte_results)
save_plot(p, "heatmap.png")
}
}
