% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_results}
\alias{save_results}
\title{Save Analysis Results}
\usage{
save_results(results, file_path)
}
\arguments{
\item{results}{List of analysis results}

\item{file_path}{File path for saving}
}
\value{
Invisible NULL
}
\description{
Save Analysis Results
}
\examples{
\dontrun{
save_results(list(qte = qte_results), "analysis_results.RData")
}
}
