% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_decomposition.R
\name{wavelet_decompose}
\alias{wavelet_decompose}
\title{Perform Wavelet Decomposition}
\usage{
wavelet_decompose(returns, n.levels = 4, wf = "la8")
}
\arguments{
\item{returns}{Numeric vector of return data}

\item{n.levels}{Number of wavelet decomposition levels (default: 4)}

\item{wf}{Wavelet filter to use (default: "la8")}
}
\value{
List containing wavelet decomposition results
}
\description{
Perform Wavelet Decomposition
}
\examples{
\donttest{
data <- rnorm(256)
wave <- wavelet_decompose(data, n.levels = 4)
}
}
