% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{process_returns}
\alias{process_returns}
\title{Process Returns Data}
\usage{
process_returns(returns)
}
\arguments{
\item{returns}{xts object containing return data}
}
\value{
Processed xts object
}
\description{
Process Returns Data
}
\examples{
\donttest{
data <- get_stock_data(data_source = "builtin")
processed_data <- process_returns(data)
}
}
